pyproject.toml
[tool.poetry]
name = "react-client-side"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.2"
SQLAlchemy = "^1.4.27"
alembic = "^1.7.5"
Flask-SQLAlchemy = "^2.5.1"
tenacity = "^8.0.1"
Flask-Cors = "^3.0.10"
Flask-JWT-Extended = "^4.3.1"
flask-apispec = "^0.11.0"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.910"
black = "^21.11b1"
isort = "^5.10.1"
pytest-cov = "^3.0.0"
flake8 = "^4.0.1"
autoflake = "^1.4"
sqlalchemy-stubs = "^0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
